requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
cbor==1.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
chonkie==1.6.0
chonkie-core==0.9.2
click==8.3.1
datasets==4.7.0
dill==0.4.0
fastapi==0.135.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
ijson==3.5.0
iniconfig==2.3.0
inscriptis==2.7.1
ir_datasets==0.5.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
peft==0.18.1
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.2
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qdrant-client==1.17.1
rank-bm25==0.2.2
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
sentencepiece==0.2.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
socksio==1.0.0
soupsieve==2.8.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
trec-car-tools==2.6
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
unlzw3==0.2.3
urllib3==2.6.3
uvicorn==0.42.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
xxhash==3.6.0
yarl==1.23.0
zlib-state==0.1.10
